app/src/main/java/com/vimal/quote/api/RetrofitService.kt
package com.vimal.quote.api

/**
 * Created by Vimal on Feb-2023.
 */

import com.vimal.quote.model.Quote
import retrofit2.Response
import retrofit2.Retrofit
import retrofit2.converter.gson.GsonConverterFactory
import retrofit2.http.GET

interface RetrofitService {

    @GET("quotes")
    suspend fun getAllQuotes() : Response<List<Quote>>

    companion object {
        var retrofitService: RetrofitService? = null
        fun getInstance() : RetrofitService {
            if (retrofitService == null) {
                val retrofit = Retrofit.Builder()
                    .baseUrl("https://type.fit/api/")
                    .addConverterFactory(GsonConverterFactory.create())
                    .build()
                retrofitService = retrofit.create(RetrofitService::class.java)
            }
            return retrofitService!!
        }

    }
}